Correct all compiler warning messages.
  Code compiled with:
gcc -Wall -Wbad-function-cast -Wdeclaration-after-statement \
    -Wmissing-prototypes -Wmissing-declarations -Wnested-externs \
    -fno-strict-aliasing -Wold-style-definition -Wpointer-arith \
    -Wstrict-prototypes

  This patch also makes most smi_501.c functions static, as they are not
called from anywhere else.
  Also removed prototypes for non existent functions in smi_501.h.

+++ src/smi_hwcurs.c
@@ -351,14 +351,14 @@ SMI_LoadCursorImage(ScrnInfoPtr pScrn, unsigned char *src)
 	    WRITE_DCR(pSmi, DCRF0, pSmi->FBCursorOffset);
 	    xf86DrvMsgVerb(pScrn->scrnIndex, X_INFO, VERBLEV,
 			   "Primary FBCursorOffset at 0x%08X\n",
-			   pSmi->FBCursorOffset);
+			   (unsigned int)pSmi->FBCursorOffset);
 	}
 	else {
 	    /* CRT   HWC Addr */
 	    WRITE_DCR(pSmi, DCR230, pSmi->videoRAMBytes + pSmi->FBCursorOffset);
 	    xf86DrvMsgVerb(pScrn->scrnIndex, X_INFO, VERBLEV,
 			   "Secondary FBCursorOffset at 0x%08X\n",
-			   pSmi->FBCursorOffset);
+			   (unsigned int)pSmi->FBCursorOffset);
 	}
     }
     else {
